Fix #268: 修复改用`shutil`移动文件后可能导致文件被覆盖的问题
Introduced in 37c2554c24734d3326d4e0f9b4faf6ac677c7453

+++ core/datatype.py
@@ -151,6 +151,9 @@ class Movie:
         def move_file(src:str, dst:str):
             """移动（重命名）文件并记录信息到日志"""
             abs_dst = os.path.abspath(dst)
+            # shutil.move might overwrite dst file
+            if os.path.exists(abs_dst):
+                raise FileExistsError(f'File exists: {abs_dst}')
             shutil.move(src, abs_dst)
             src_rel = os.path.relpath(src)
             dst_name = os.path.basename(dst)
